// libraries/googleapis/protos/google/type/decimal_pb.d.ts
// @generated by protoc-gen-es v2.2.3 with parameter "keep_empty_files=true,target=js+dts,import_extension=js"
// @generated from file libraries/googleapis/protos/google/type/decimal.proto (package google.type, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file libraries/googleapis/protos/google/type/decimal.proto.
 */
export declare const file_libraries_googleapis_protos_google_type_decimal: GenFile;

/**
 * @generated from message google.type.Decimal
 */
export declare type Decimal = Message<"google.type.Decimal"> & {
  /**
   * @generated from field: string value = 1;
   */
  value: string;
};

/**
 * Describes the message google.type.Decimal.
 * Use `create(DecimalSchema)` to create a new message.
 */
export declare const DecimalSchema: GenMessage<Decimal>;